+++ CHANGELOG.md
@@ -55,6 +55,7 @@ All notable changes to this project will be documented in this file.
 - Marked OxyPlot.Xamarin.Forms.Platform.*.Forms.Init() obsolete (#632)
 - Throw exception if Xamarin.Forms renderer is not 'initialized' (#492)
 - Make numeric values of DateTimeAxis compatible with ToOADate (#660)
+- Make DataPoint immutable (#692)
 
 ### Removed
 - StyleCop tasks (#556)